tests/requirements-test.txt
# Test Requirements
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests-mock>=1.9.0
httpx>=0.24.0
fastapi[all]>=0.68.0

# For mocking
pytest-mock>=3.6.0

# For coverage reporting (optional)
pytest-cov>=4.0.0
